remove c# devkit from devcontainer
i don't like it enough, and it's non-free. the c# extension is not free
as in speech either, but at least it's free as in beer.

i was attempting to use it particularly for the test explorer, but it
wasn't working well enough for me to justify keeping it around.