make tasks for running all the tests under watch
with this setup the dev can activate the "run test task" action and
it will start the "watch all tests" -task.

the watch all tests task starts each of the test projects in a separate
terminal (side-by-side with each other).

these watch test -tasks watch the tests and will mark problems in the
problems panel (and in the code on the correct line). they also collect
coverage to files that the coverage-gutter monitors.

with this setup the dev can see the test results and coverage in real
time while developing.